🔧 Extract sample to own Gradle project
