+++ build.gradle.kts
@@ -1,17 +1,12 @@
 buildscript {
     repositories {
         mavenCentral()
-        maven("https://nexus.eman.cz/repository/maven-public")
-    }
-
-    dependencies {
-        classpath("cz.eman.swagger:swagger-codegen:${libs.versions.swaggerCodegen.get()}")
     }
 }
 
 plugins {
     alias(libs.plugins.kotlin.jvm) apply false
-    alias(libs.plugins.kotlin.kapt) apply false
+    alias(libs.plugins.dokka) apply false
 }
 
 allprojects {
