+++ sample/gradle/libs.versions.toml
@@ -1,0 +1,27 @@
+[versions]
+gradle = "8.4"
+
+kotlin = "1.9.10"
+
+moshi = "1.15.0"
+retrofit = "2.9.0"
+
+
+[plugins]
+kotlin-jvm = { id = "org.jetbrains.kotlin.jvm", version.ref = "kotlin" }
+kotlin-kapt = { id = "org.jetbrains.kotlin.kapt", version.ref = "kotlin" }
+
+
+[libraries]
+moshi-core = { group = "com.squareup.moshi", name = "moshi", version.ref = "moshi" }
+moshi-adapters = { group = "com.squareup.moshi", name = "moshi-adapters", version.ref = "moshi" }
+moshi-codegen = { group = "com.squareup.moshi", name = "moshi-kotlin-codegen", version.ref = "moshi" }
+
+retrofit-core = { group = "com.squareup.retrofit2", name = "retrofit", version.ref = "retrofit" }
+retrofit-converter-scalars = { group = "com.squareup.retrofit2", name = "converter-scalars", version.ref = "retrofit" }
+retrofit-converter-moshi = { group = "com.squareup.retrofit2", name = "converter-moshi", version.ref = "retrofit" }
+
+
+[bundles]
+moshi = ["moshi-core", "moshi-adapters"]
+retrofit = ["retrofit-core", "retrofit-converter-scalars", "retrofit-converter-moshi"]
