+++ sample/kotlin-client/openapi/one-of-animal1.yaml
