+++ sample/kotlin-client/openapi/one-of-animal2.yaml
