+++ sample/kotlin-client/openapi/one-of.yaml
