+++ sample/kotlin-client/openapi/petstore-opt-def-arg-api.yaml
