+++ sample/kotlin-client/openapi/petstore.yaml
