+++ settings.gradle.kts
@@ -1,4 +1,3 @@
 rootProject.buildFileName = "build.gradle.kts"
 
-include(":lib", ":sample")
-
+include(":lib")
